+++ parxe/engines/pbs.lua
@@ -60,7 +60,12 @@ function pbs:destructor()
     self.server:shutdown(self.endpoint)
     self.server:close()
   end
-  os.remove(self.TMPNAME)
+  if self.TMPNAME then os.remove(self.TMPNAME) end
+end
+
+function pbs_methods:destroy()
+  pbs.destructor(self)
+  for k,v in pairs(self) do self[k] = nil end
 end
 
 function pbs_methods:init()
