+++ parxe/engines/ssh.lua
@@ -65,7 +65,12 @@ function ssh:destructor()
     self.server:shutdown(self.endpoint)
     self.server:close()
   end
-  os.remove(self.TMPNAME)
+  if self.TMPNAME then os.remove(self.TMPNAME) end
+end
+
+function ssh_methods:destroy()
+  ssh.destructor(self)
+  for k,v in pairs(self) do self[k] = nil end
 end
 
 function ssh_methods:init()
